requirements.txt
streamlit==1.19.0
openai==0.27.0
faiss-cpu==1.7.2
pandas==1.5.3